changed attach functions to internal
These functions should be a part of the implementation only. Putting them in the interface as external makes them less secure. The standard should be a secure building block, if it isn't secure there is no reason to use it. If these functions are needed externally so an inheriting contract can call them.